pyproject.toml
[tool.poetry]
name = "habit-api"
version = "0.1.0"
description = ""
authors = ["Cosmic Onyx <c0sm1c0nyx.225@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.8"
djangorestframework = "^3.16.1"
celery = "^5.5.3"
djangorestframework-simplejwt = "^5.5.1"
python-dotenv = "^1.2.1"
drf-nested-routers = "^0.95.0"
django-extensions = "^4.1"
loguru = "^0.7.3"
notifiers = "^1.3.6"
django-filter = "^25.2"
psycopg2-binary = "^2.9.11"
redis = "^7.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
